refactor: move stages to combat package and remove New() functions
- Move StageBase, StageFeatures, etc. from dnd5e to combat package
  (they are combat chain concerns)
- Delete dnd5e/stages.go entirely (no backward compat re-exports)
- Update conditions and combat to use combat.StageFeatures
- Remove New() and NewInput from features and conditions packages
  (postpone initialization design decision - YAGNI)
- Keep Type constants, ID constants, and Grant structs (valuable)

This breaks the import cycle that prevented classes from importing
features/conditions, enabling future level-based grant definitions.
