+++ rulebooks/dnd5e/combat/stages.go
@@ -1,7 +1,7 @@
 // Copyright (C) 2024 Kirk Diggler
 // SPDX-License-Identifier: GPL-3.0-or-later
 
-package dnd5e
+package combat
 
 import "github.com/KirkDiggler/rpg-toolkit/core/chain"
